Set active arrays in vtk polydata
Previously untrakofy output a vtkPolyData with arrays
for each of the restored attributes, but it did not
set any of them active.  Some SlicerDMRI utilities require
a default active tensor be defined (even though it's not
semantically meaningful).

This change sets the first scalar array and the first
tensor array to be active in order to be compatible
with the expectations of the FiberTractMeasurements
tool in SlicerDMRI.

Fixes #11